backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
pydantic
python-dotenv
python-multipart
aiofiles
boto3
motor[srv]
passlib[bcrypt]
python-jose[cryptography]
PyJWT
email-validator